# src/swagger/prd-api-2-0.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Motor Cards V3
  description: Motor de Cards Natura - v3
  contact: {}
host: "motor-cards-v3.endpoints.natura-191517.cloud.goog"
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/jobs/check/{id}':
    get:
      security:
      - api_key: []
      description: Checks if a job has been completed
      summary: Checks if a job has been completed
      tags:
        - Jobs
      operationId: CheckStatusJob
      produces:
        - application/json
      parameters:
        - name: userid
          in: header
          required: true
          type: string
          description: ''
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: Get with success
          schema:
            example: 
              status: 
               success: true
               message: "DONE"
        '412':
          description: Get without userid header
          schema: 
            example: 
              status: 
               success: false
               message: "You aren't logged"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"